KNN ( K Nearest Neighbours)
KNN is a supervised machine learning algorithm. It can be used to classification of data set. Here I will be sharing code for implementing kNN and share some tricks for KNN. The dataset contains the Age, Salary and the decision to purchase a particular Car. We will be using KNN to predict if a particular person will buy a car. The notebook will cover following topics

1.Data Import and Preprocessing

2.Exploratory Data Analysis

3.Model Built

4.Model Evaluation

5.Optimum Value of k

6.Conclusion